Time: 12 ms (25.11%), Space: 13.9 MB (17.29%) - LeetHub

+++ 152-maximum-product-subarray/152-maximum-product-subarray.cpp
@@ -4,7 +4,7 @@ public:
         int large = nums.front();
         int small = nums.front();
         int result = large;
-        for(int i = 1; i < nums.size();i++){
+        for(short i = 1; i < nums.size();i++){
             int small_copy = small;
             
             small = min(min(nums[i], small * nums[i]),large * nums[i]);
